# ggml : synchronize threads using barriers