Replaces the spinlocks in the CPU backend with barriers, using openmp if possible.

| CPU                                  | Model                      |   Threads | Test   |   t/s master |   t/s sl/test-omp-sync |   Speedup |
|:-------------------------------------|:---------------------------|----------:|:-------|-------------:|-----------------------:|----------:|
| i9-13900K | stories260k F32 |         8 | pp512  |     37153.86 |               34901.28 |      0.94 |
| i9-13900K | stories260k F32 |         8 | tg128  |      2926.66 |                4913.27 |      1.68 |
| i9-13900K | stories260k F32 |        16 | pp512  |     20069.17 |               21167.47 |      1.05 |
| i9-13900K | stories260k F32 |        16 | tg128  |      1450.93 |                3405.28 |      2.35 |
| i9-13900K | stories260k F32 |        24 | pp512  |     24257.18 |               26325.85 |      1.09 |
| i9-13900K | stories260k F32 |        24 | tg128  |      1022.16 |                2734.91 |      2.68 |
| i9-13900K |  1B Q8_0 |         8 | pp128  |       340.93 |                 336.43 |      0.99 |
| i9-13900K |  1B Q8_0 |         8 | tg32   |        70.62 |                  73.01 |      1.03 |
| i9-13900K |  1B Q8_0 |        16 | pp128  |       437.92 |                 424.76 |      0.97 |
| i9-13900K |  1B Q8_0 |        16 | tg32   |        58.12 |                  64.29 |      1.11 |
| i9-13900K |  1B Q8_0 |        24 | pp128  |       521.29 |                 506.86 |      0.97 |
| i9-13900K |  1B Q8_0 |        24 | tg32   |        53.88 |                  63.47 |      1.18 |
| i9-13900K | 7B Q4_0 |         8 | pp128  |        49.77 |                  49.81 |      1.00 |
| i9-13900K | 7B Q4_0 |         8 | tg32   |        21.07 |                  21.41 |      1.02 |
| i9-13900K | 7B Q4_0 |        16 | pp128  |        63.52 |                  64.56 |      1.02 |
| i9-13900K | 7B Q4_0 |        16 | tg32   |        20.71 |                  21.93 |      1.06 |
| i9-13900K | 7B Q4_0 |        24 | pp128  |        73.48 |                  73.90 |      1.01 |
| i9-13900K | 7B Q4_0 |        24 | tg32   |        17.93 |                  19.24 |      1.07 |