@Green-Sky 2024-06-18
> spinlocks

Is there it is not using mtecies ? (mutexes? o.o)