@slaren 2024-06-18
Since these waits are often very short, spinlocks resulted in better performance. OpenMP is also likely spinning for a while before sleeping.