@slaren 2024-06-18
There were several PRs to change the synchronization mechanism in the past, but always resulted in worse performance.