@slaren 2024-06-18
The fallback seems to perform ok on the M3 Max.

`LLAMA_NO_METAL=1 LLAMA_NO_ACCELERATE=1 LLAMA_NO_LLAMAFILE=1`

| CPU   | Model                      |   Threads | Test   |   t/s master |   t/s sl/test-omp-sync |   Speedup |
|:------|:---------------------------|----------:|:-------|-------------:|-----------------------:|----------:|
| M3 Max | stories260k F32 |         4 | pp512  |     24415.91 |               24354.48 |      1.00 |
| M3 Max | stories260k F32 |         4 | tg128  |      6134.50 |                7296.25 |      1.19 |
| M3 Max | stories260k F32 |         8 | pp512  |     40683.65 |               41209.17 |      1.01 |
| M3 Max | stories260k F32 |         8 | tg128  |      2503.32 |                3272.53 |      1.31 |
| M3 Max | stories260k F32 |        12 | pp512  |     55364.55 |               55204.87 |      1.00 |
| M3 Max | stories260k F32 |        12 | tg128  |      1977.22 |                2406.93 |      1.22 |
| M3 Max | stories260k F32 |        16 | pp512  |     44830.02 |               49383.96 |      1.10 |
| M3 Max | stories260k F32 |        16 | tg128  |      1058.14 |                1436.44 |      1.36 |
| M3 Max | 1B Q8_0 |         4 | pp128  |       201.11 |                 190.80 |      0.95 |
| M3 Max | 1B Q8_0 |         4 | tg64   |       100.37 |                 101.50 |      1.01 |
| M3 Max | 1B Q8_0 |         8 | pp128  |       359.73 |                 353.41 |      0.98 |
| M3 Max | 1B Q8_0 |         8 | tg64   |        95.07 |                  99.61 |      1.05 |
| M3 Max | 1B Q8_0 |        12 | pp128  |       425.86 |                 455.97 |      1.07 |
| M3 Max | 1B Q8_0 |        12 | tg64   |        93.11 |                  97.74 |      1.05 |
| M3 Max | 1B Q8_0 |        16 | pp128  |       408.75 |                 494.34 |      1.21 |
| M3 Max | 1B Q8_0 |        16 | tg64   |        26.22 |                  59.93 |      2.29 |
| M3 Max | 7B Q4_0 |         4 | pp128  |        25.15 |                  24.98 |      0.99 |
| M3 Max | 7B Q4_0 |         4 | tg64   |        22.10 |                  22.08 |      1.00 |
| M3 Max | 7B Q4_0 |         8 | pp128  |        46.31 |                  46.11 |      1.00 |
| M3 Max | 7B Q4_0 |         8 | tg64   |        29.73 |                  30.26 |      1.02 |
| M3 Max | 7B Q4_0 |        12 | pp128  |        58.64 |                  58.82 |      1.00 |
| M3 Max | 7B Q4_0 |        12 | tg64   |        29.12 |                  29.91 |      1.03 |
| M3 Max | 7B Q4_0 |        16 | pp128  |        65.84 |                  67.42 |      1.02 |
| M3 Max | 7B Q4_0 |        16 | tg64   |         6.70 |                  24.69 |      3.69 |
| M3 Max | 7B F32 |         4 | pp128  |         9.65 |                   9.66 |      1.00 |
| M3 Max | 7B F32 |         4 | tg64   |         4.75 |                   4.77 |      1.00 |
| M3 Max | 7B F32 |         8 | pp128  |        17.40 |                  17.87 |      1.03 |
| M3 Max | 7B F32 |         8 | tg64   |         4.75 |                   4.77 |      1.00 |
| M3 Max | 7B F32 |        12 | pp128  |        22.86 |                  22.93 |      1.00 |
| M3 Max | 7B F32 |        12 | tg64   |         4.70 |                   4.74 |      1.01 |
| M3 Max | 7B F32 |        16 | pp128  |        23.05 |                  24.70 |      1.07 |
| M3 Max | 7B F32 |        16 | tg64   |         0.77 |                   4.62 |      5.99 |