@slaren 2024-06-18
The server thread sanitizer run failed the [first time](https://github.com/ggerganov/llama.cpp/actions/runs/9570822053/job/26386465576?pr=7993), but I don't understand the error, I cannot reproduce it locally, and the second run passed. It may have been a fluke, but it's something to look at if it continues failing intermittently after this is merged.