@slaren 2024-06-18
With MSVC:
| CPU   | Model           |   Threads | Test   |   t/s master |   t/s sl/test-omp-sync |   Speedup |
|:------|:----------------|----------:|:-------|-------------:|-----------------------:|----------:|
| i9-13900K | stories260k F32 |         8 | pp512  |     29900.69 |               29218.33 |      0.98 |
| i9-13900K | stories260k F32 |         8 | tg128  |      2168.78 |                3604.36 |      1.66 |
| i9-13900K | stories260k F32 |        16 | pp512  |     14942.76 |               15333.24 |      1.03 |
| i9-13900K | stories260k F32 |        16 | tg128  |       651.22 |                2253.29 |      3.46 |
| i9-13900K | stories260k F32 |        24 | pp512  |     18434.45 |               19764.31 |      1.07 |
| i9-13900K | stories260k F32 |        24 | tg128  |       418.43 |                1637.96 |      3.91 |
| i9-13900K | 1B Q8_0         |         8 | pp128  |       298.91 |                 304.23 |      1.02 |
| i9-13900K | 1B Q8_0         |         8 | tg64   |        60.63 |                  65.54 |      1.08 |
| i9-13900K | 1B Q8_0         |        16 | pp128  |       254.52 |                 257.34 |      1.01 |
| i9-13900K | 1B Q8_0         |        16 | tg64   |        43.94 |                  58.29 |      1.33 |
| i9-13900K | 1B Q8_0         |        24 | pp128  |       355.98 |                 368.96 |      1.04 |
| i9-13900K | 1B Q8_0         |        24 | tg64   |        41.27 |                  64.36 |      1.56 |
| i9-13900K | 7B Q4_0         |         8 | pp128  |        37.51 |                  38.40 |      1.02 |
| i9-13900K | 7B Q4_0         |         8 | tg64   |        15.62 |                  15.96 |      1.02 |
| i9-13900K | 7B Q4_0         |        16 | pp128  |        32.53 |                  32.61 |      1.00 |
| i9-13900K | 7B Q4_0         |        16 | tg64   |        11.88 |                  12.99 |      1.09 |
| i9-13900K | 7B Q4_0         |        24 | pp128  |        47.63 |                  48.02 |      1.01 |
| i9-13900K | 7B Q4_0         |        24 | tg64   |        12.60 |                  15.55 |      1.23 |