@ggerganov 2024-06-19
Results on M1 Pro and M2 Ultra:

```bash
LLAMA_NO_METAL=1 LLAMA_NO_ACCELERATE=1 LLAMA_NO_LLAMAFILE=1 ./scripts/compare-commits.sh master sl/test-omp-sync -m models/tinyllama-1b-chat/ggml-model-f16.gguf -m models/tinyllama-1b-chat/ggml-model-q8_0.gguf -m models/tinyllama-1b-chat/ggml-model-q4_0.gguf -p 128 -n 64 -t 4,8
```

| CPU   | Model         |   [GiB] |   Threads | Test   |   t/s master |   t/s sl/test-omp-sync |   Speedup |
|:------|:--------------|--------:|----------:|:-------|-------------:|-----------------------:|----------:|
| M1 Pro | llama 1B F16  |    2.05 |         4 | pp128  |       113.37 |                 109.21 |      0.96 |
| M1 Pro | llama 1B F16  |    2.05 |         4 | tg64   |        44.83 |                  45.88 |      1.02 |
| M1 Pro | llama 1B F16  |    2.05 |         8 | pp128  |       204.39 |                 172.73 |      0.85 |
| M1 Pro | llama 1B F16  |    2.05 |         8 | tg64   |        38.98 |                  49.44 |      1.27 |
| M1 Pro | llama 1B Q4_0 |    0.59 |         4 | pp128  |       128.83 |                 123.33 |      0.96 |
| M1 Pro | llama 1B Q4_0 |    0.59 |         4 | tg64   |        90.91 |                  90.20 |      0.99 |
| M1 Pro | llama 1B Q4_0 |    0.59 |         8 | pp128  |       206.68 |                 203.71 |      0.99 |
| M1 Pro | llama 1B Q4_0 |    0.59 |         8 | tg64   |       117.38 |                 122.94 |      1.05 |
| M1 Pro | llama 1B Q8_0 |    1.09 |         4 | pp128  |       160.17 |                 143.42 |      0.90 |
| M1 Pro | llama 1B Q8_0 |    1.09 |         4 | tg64   |        71.12 |                  71.98 |      1.01 |
| M1 Pro | llama 1B Q8_0 |    1.09 |         8 | pp128  |       257.38 |                 237.44 |      0.92 |
| M1 Pro | llama 1B Q8_0 |    1.09 |         8 | tg64   |        60.25 |                  84.80 |      1.41 |

| CPU   | Model         |    [GiB] |   Threads | Test   |   t/s master |   t/s sl/test-omp-sync |   Speedup |
|:------|:--------------|---------:|----------:|:-------|-------------:|-----------------------:|----------:|
| M2 Ultra | llama 1B F16  |     2.05 |         4 | pp128  |       118.08 |                 121.59 |      1.03 |
| M2 Ultra | llama 1B F16  |     2.05 |         4 | tg64   |        58.73 |                  53.21 |      0.91 |
| M2 Ultra | llama 1B F16  |     2.05 |         8 | pp128  |       238.14 |                 239.27 |      1.00 |
| M2 Ultra | llama 1B F16  |     2.05 |         8 | tg64   |        66.46 |                  69.03 |      1.04 |
| M2 Ultra | llama 1B F16  |     2.05 |        16 | pp128  |       439.46 |                 432.57 |      0.98 |
| M2 Ultra | llama 1B F16  |     2.05 |        16 | tg64   |        72.97 |                  82.59 |      1.13 |
| M2 Ultra | llama 1B Q4_0 |     0.59 |         4 | pp128  |       147.08 |                 147.61 |      1.00 |
| M2 Ultra | llama 1B Q4_0 |     0.59 |         4 | tg64   |        90.74 |                  91.91 |      1.01 |
| M2 Ultra | llama 1B Q4_0 |     0.59 |         8 | pp128  |       285.33 |                 285.09 |      1.00 |
| M2 Ultra | llama 1B Q4_0 |     0.59 |         8 | tg64   |       136.46 |                 145.17 |      1.06 |
| M2 Ultra | llama 1B Q4_0 |     0.59 |        16 | pp128  |       526.63 |                 526.64 |      1.00 |
| M2 Ultra | llama 1B Q4_0 |     0.59 |        16 | tg64   |       136.58 |                 164.73 |      1.21 |
| M2 Ultra | llama 1B Q8_0 |     1.09 |         4 | pp128  |       174.99 |                 175.82 |      1.00 |
| M2 Ultra | llama 1B Q8_0 |     1.09 |         4 | tg64   |        76.69 |                  80.51 |      1.05 |
| M2 Ultra | llama 1B Q8_0 |     1.09 |         8 | pp128  |       336.11 |                 336.49 |      1.00 |
| M2 Ultra | llama 1B Q8_0 |     1.09 |         8 | tg64   |        92.01 |                  98.79 |      1.07 |
| M2 Ultra | llama 1B Q8_0 |     1.09 |        16 | pp128  |       595.29 |                 612.25 |      1.03 |
| M2 Ultra | llama 1B Q8_0 |     1.09 |        16 | tg64   |       108.28 |                 123.60 |      1.14 |


M2 Ultra seems to respond better to the changes