@ggerganov 2024-06-19
After removing `sched_yield`, here are the results:

| CPU   | Model         |   [GiB] |   Threads | Test   |   t/s master |   t/s sl/test-omp-sync |   Speedup |
|:------|:--------------|--------:|----------:|:-------|-------------:|-----------------------:|----------:|
| M1 Pro | llama 1B F16  |    2.05 |         4 | pp128  |       113.82 |                 113.25 |      0.99 |
| M1 Pro | llama 1B F16  |    2.05 |         4 | tg64   |        44.87 |                  45.90 |      1.02 |
| M1 Pro | llama 1B F16  |    2.05 |         8 | pp128  |       220.78 |                 188.19 |      0.85 |
| M1 Pro | llama 1B F16  |    2.05 |         8 | tg64   |        38.35 |                  39.43 |      1.03 |
| M1 Pro | llama 1B Q4_0 |    0.59 |         4 | pp128  |       136.69 |                 126.87 |      0.93 |
| M1 Pro | llama 1B Q4_0 |    0.59 |         4 | tg64   |        91.48 |                  92.21 |      1.01 |
| M1 Pro | llama 1B Q4_0 |    0.59 |         8 | pp128  |       230.98 |                 209.10 |      0.91 |
| M1 Pro | llama 1B Q4_0 |    0.59 |         8 | tg64   |        85.29 |                 124.21 |      1.46 |
| M1 Pro | llama 1B Q8_0 |    1.09 |         4 | pp128  |       164.59 |                 153.72 |      0.93 |
| M1 Pro | llama 1B Q8_0 |    1.09 |         4 | tg64   |        70.59 |                  72.08 |      1.02 |
| M1 Pro | llama 1B Q8_0 |    1.09 |         8 | pp128  |       289.09 |                 251.00 |      0.87 |
| M1 Pro | llama 1B Q8_0 |    1.09 |         8 | tg64   |        55.05 |                  57.00 |      1.04 |